fix: carousel
