+++ src/assets/images/carousel/index.tsx
@@ -1,0 +1,22 @@
+import Library1 from './library1.jpg';
+import Library2 from './library2.jpg';
+import Library3 from './library3.jpg';
+import Library4 from './library4.jpg';
+import Library5 from './library5.jpg';
+
+import Carousel1 from './carousel1.jpg';
+import Carousel2 from './carousel2.jpg';
+import Carousel3 from './carousel3.jpg';
+import Carousel4 from './carousel4.jpg';
+
+export {
+  Carousel1,
+  Carousel2,
+  Carousel3,
+  Carousel4,
+  Library1,
+  Library2,
+  Library3,
+  Library4,
+  Library5
+};
