--- src/image/index.tsx
@@ -1,7 +1,0 @@
-import Library1 from './library1.jpg';
-import Library2 from './library2.jpg';
-import Library3 from './library3.jpg';
-import Library4 from './library4.jpg';
-import Library5 from './library5.jpg';
-
-export { Library1, Library2, Library3, Library4, Library5 };
